test(interp): simplify test ownership and structure
