+++ interp/trace_test.go
@@ -170,77 +170,73 @@ func TestTracer_Capture(t *testing.T) {
 		require.Equal(t, prof.CaptureReasonUnsupportedOp, result.reason)
 	})
 
-	t.Run("isolates captured mutations", func(t *testing.T) {
-		primitive := types.TypedArray[int32]{1}
-		array := types.NewArray(types.TypeI32Array, types.BoxI32(1))
-		structure := types.NewStruct(
-			types.NewStructType(types.NewStructField(types.TypeI32)),
-			types.BoxI32(1),
-		)
-		tests := []struct {
-			name  string
-			value types.Value
-			op    instr.Opcode
-			read  func() types.Boxed
-		}{
-			{name: "primitive array", value: primitive, op: instr.ARRAY_SET, read: func() types.Boxed { return types.BoxI32(primitive[0]) }},
-			{name: "boxed array", value: array, op: instr.ARRAY_SET, read: func() types.Boxed { return array.Elems[0] }},
-			{name: "struct", value: structure, op: instr.STRUCT_SET, read: func() types.Boxed { return structure.Field(0) }},
-		}
-		for _, tt := range tests {
-			t.Run(tt.name, func(t *testing.T) {
-				tracer := NewTracer()
-				prog := program.New([]instr.Instruction{
-					instr.New(instr.CONST_GET, 0),
-					instr.New(instr.I32_CONST, 0),
-					instr.New(instr.I32_CONST, 2),
-					instr.New(tt.op),
-				}, program.WithConstants(tt.value))
-				i := New(prog, WithTracer(tracer), WithThreshold(-1))
-				defer i.Close()
+	primitive := types.TypedArray[int32]{1}
+	array := types.NewArray(types.TypeI32Array, types.BoxI32(1))
+	structure := types.NewStruct(
+		types.NewStructType(types.NewStructField(types.TypeI32)),
+		types.BoxI32(1),
+	)
+	mutationTests := []struct {
+		name  string
+		value types.Value
+		op    instr.Opcode
+		read  func() types.Boxed
+	}{
+		{name: "primitive array", value: primitive, op: instr.ARRAY_SET, read: func() types.Boxed { return types.BoxI32(primitive[0]) }},
+		{name: "boxed array", value: array, op: instr.ARRAY_SET, read: func() types.Boxed { return array.Elems[0] }},
+		{name: "struct", value: structure, op: instr.STRUCT_SET, read: func() types.Boxed { return structure.Field(0) }},
+	}
+	for _, tt := range mutationTests {
+		t.Run("isolates captured mutation for "+tt.name, func(t *testing.T) {
+			tracer := NewTracer()
+			prog := program.New([]instr.Instruction{
+				instr.New(instr.CONST_GET, 0),
+				instr.New(instr.I32_CONST, 0),
+				instr.New(instr.I32_CONST, 2),
+				instr.New(tt.op),
+			}, program.WithConstants(tt.value))
+			i := New(prog, WithTracer(tracer), WithThreshold(-1))
+			defer i.Close()
 
-				tracer.capture(i, anchor{})
-				require.Equal(t, types.BoxI32(1), tt.read())
-			})
-		}
-	})
+			tracer.capture(i, anchor{})
+			require.Equal(t, types.BoxI32(1), tt.read())
+		})
+	}
 
-	t.Run("preserves typed array aliases", func(t *testing.T) {
-		shared := types.TypedArray[int32]{0}
-		backing := types.TypedArray[int32]{0, 0, 0}
-		tests := []struct {
-			name      string
-			first     types.TypedArray[int32]
-			second    types.TypedArray[int32]
-			setIndex  uint64
-			readIndex uint64
-			original  func() int32
-		}{
-			{name: "same slice", first: shared, second: shared, original: func() int32 { return shared[0] }},
-			{name: "overlapping subslices", first: backing[:2:2], second: backing[1:3], setIndex: 1, original: func() int32 { return backing[1] }},
-		}
-		for _, tt := range tests {
-			t.Run(tt.name, func(t *testing.T) {
-				tracer := NewTracer()
-				prog := program.New([]instr.Instruction{
-					instr.New(instr.CONST_GET, 0),
-					instr.New(instr.I32_CONST, tt.setIndex),
-					instr.New(instr.I32_CONST, 1),
-					instr.New(instr.ARRAY_SET),
-					instr.New(instr.CONST_GET, 1),
-					instr.New(instr.I32_CONST, tt.readIndex),
-					instr.New(instr.ARRAY_GET),
-				}, program.WithConstants(tt.first, tt.second))
-				i := New(prog, WithTracer(tracer), WithThreshold(-1))
-				defer i.Close()
+	shared := types.TypedArray[int32]{0}
+	backing := types.TypedArray[int32]{0, 0, 0}
+	aliasTests := []struct {
+		name      string
+		first     types.TypedArray[int32]
+		second    types.TypedArray[int32]
+		setIndex  uint64
+		readIndex uint64
+		original  func() int32
+	}{
+		{name: "same slice", first: shared, second: shared, original: func() int32 { return shared[0] }},
+		{name: "overlapping subslices", first: backing[:2:2], second: backing[1:3], setIndex: 1, original: func() int32 { return backing[1] }},
+	}
+	for _, tt := range aliasTests {
+		t.Run("preserves typed array aliases for "+tt.name, func(t *testing.T) {
+			tracer := NewTracer()
+			prog := program.New([]instr.Instruction{
+				instr.New(instr.CONST_GET, 0),
+				instr.New(instr.I32_CONST, tt.setIndex),
+				instr.New(instr.I32_CONST, 1),
+				instr.New(instr.ARRAY_SET),
+				instr.New(instr.CONST_GET, 1),
+				instr.New(instr.I32_CONST, tt.readIndex),
+				instr.New(instr.ARRAY_GET),
+			}, program.WithConstants(tt.first, tt.second))
+			i := New(prog, WithTracer(tracer), WithThreshold(-1))
+			defer i.Close()
 
-				result := tracer.capture(i, anchor{})
-				require.NotNil(t, result.trace)
-				require.Equal(t, types.BoxI32(1), result.trace.ops[len(result.trace.ops)-1].seen)
-				require.Zero(t, tt.original())
-			})
-		}
-	})
+			result := tracer.capture(i, anchor{})
+			require.NotNil(t, result.trace)
+			require.Equal(t, types.BoxI32(1), result.trace.ops[len(result.trace.ops)-1].seen)
+			require.Zero(t, tt.original())
+		})
+	}
 
 	t.Run("cuts an oversized trace at a resumable boundary", func(t *testing.T) {
 		code := make([]instr.Instruction, opLimit+1)
@@ -389,7 +385,7 @@ func TestTracer_Capture(t *testing.T) {
 		select {
 		case <-done:
 		case <-time.After(time.Second):
-			t.Fatal("capture deadlocked while reclaiming a function")
+			require.Fail(t, "capture deadlocked while reclaiming a function")
 		}
 		require.NotNil(t, tracer.rootAt(anchor{}))
 		require.NotNil(t, tracer.rootAt(root))
